refactor(homeComponents) : Moved home page components to seprate directory
