+++ src/modules/home/ui/components/Footer.tsx
