+++ src/modules/home/ui/components/SearchFilters/CategoriesSidebar.tsx
