+++ src/modules/home/ui/components/SearchFilters/SearchCategories.tsx
