+++ src/modules/home/ui/components/SearchFilters/SearchCategory.tsx
